requirements.txt
labelme
imgviz
pyqt5
opencv-python
pillow
numpy
yaml
argparse